+++ Packages/KeepoCore/Sources/KeepoCore/PendingOverlay.swift
@@ -11,17 +11,17 @@ public enum PendingOp: Sendable {
     /// A single leg's effect on one account — already signed. An ordinary
     /// create is one of these; an edit becomes two (subtract the old
     /// effect, add the new one); a delete is one with the amount negated.
-    case transactionDelta(accountId: UUID, amount: Decimal)
+    case transactionDelta(accountId: UUID, amount: Int64)
     /// A transfer's two legs together — `fromAmount`/`toAmount` already
     /// signed (negative out of `fromAccountId`, positive into
     /// `toAccountId`), matching what `create_transfer` itself applies.
-    case transferDelta(fromAccountId: UUID, fromAmount: Decimal, toAccountId: UUID, toAmount: Decimal)
+    case transferDelta(fromAccountId: UUID, fromAmount: Int64, toAccountId: UUID, toAmount: Int64)
     /// `set_account_balance`: not a delta — the account's balance becomes
     /// exactly `newBalance` at this point in the queue, with only ops
     /// *after* it still additive on top. Matches the RPC's own semantics:
     /// it recomputes the gap fresh against whatever the balance actually
     /// is when it finally runs, not whatever the client guessed offline.
-    case accountBalanceOverride(accountId: UUID, newBalance: Decimal)
+    case accountBalanceOverride(accountId: UUID, newBalance: Int64)
 }
 
 /// Applies queued-but-not-yet-synced writes on top of the last known
@@ -33,11 +33,11 @@ public enum PendingOverlay {
     /// "everything before this is superseded" when folded in that order.
     /// An account with no cached balance and no pending op touching it is
     /// simply absent from the result, same as it was absent from `cached`.
-    public static func accountBalances(cached: [UUID: Decimal], ops: [PendingOp]) -> [UUID: Decimal] {
-        var overrideBaseline: [UUID: Decimal] = [:]
-        var runningDelta: [UUID: Decimal] = [:]
+    public static func accountBalances(cached: [UUID: Int64], ops: [PendingOp]) -> [UUID: Int64] {
+        var overrideBaseline: [UUID: Int64] = [:]
+        var runningDelta: [UUID: Int64] = [:]
 
-        func applyDelta(_ accountId: UUID, _ amount: Decimal) {
+        func applyDelta(_ accountId: UUID, _ amount: Int64) {
             runningDelta[accountId, default: 0] += amount
         }
